Fixed issues in entities and repositories

+++ src/main/java/main/PackageRepo.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PackageRepo extends CrudRepository<Package, Integer> {
+public interface PackageRepo extends CrudRepository<Package, Long> {
 
 	// Spring Data JPA also allows you to define other query methods by simply declaring their method signature.
-    public List<Package> getPackagesById(int packageId);
+    public List<Package> getPackagesById(Long packageId);
 }
